% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{damage_simulation}
\alias{damage_simulation}
\title{Simulated damage for an earthquake event in Haiti.}
\format{
A dataframe with 350 rows and 6 variables:
\describe{
\item{building_cat}{Building category, 1 for unreinforced block walls and 2 for stone masonry}
\item{CDF}{Central damage factor corresponding to the damage grade: 0 for Grade 1, 0.5 for Grade 2, 5 for Grade 3, 20 for Grade 4, 45 for Grade 5, 80 for Grade 6, 100 for Grade 7}
\item{logPGA}{Log-transformed peak ground acceleration (PGA) where PGA is given in \%g}
\item{PGA}{PGA in \%g}
\item{Easting}{Easting coordinate of the building}
\item{Northing}{Northing coordinate of the building}
}
}
\source{
Simulated from a spatial ordinal model using logPGA simulations from OpenQuake.
}
\usage{
damage_simulation
}
\description{
A dataset containing a simulation of damage information for 350 buildings in Haiti (175 unreinforced block walls and 175 stone masonry buildings in Haiti with 25 buildings in each damage grade in the damage scale used in ATC-13).
}
\keyword{datasets}
