% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_spatial_fit}
\alias{demo_spatial_fit}
\title{Spatial ordinal model fit to the simulated damage for an earthquake event in Haiti.}
\format{
A list with 10 variables:
\describe{
\item{par}{A vector containing the estimates for for the model parameters: \code{c_factor1} (first cut-off value and cut-off ratios for Building Category 1), \code{c_factor2} (first cut-off value and cut-off ratios for Building Category 2), \code{log_phi} (log-transformed range parameter for the shared spatial field), \code{log_sigma_2} (log-transformed partial sill parameter for the shared spatial field), \code{log_phi1} (log-transformed range parameter for the Building Category 1 spatial field), \code{log_tau1_2} (log-transformed nugget parameter for the Building Category 1 spatial field), \code{log_sigma1_2} (log-transformed partial sill parameter for the Building Category 1 spatial field), \code{log_phi2} (log-transformed range parameter for the Building Category 2 spatial field), \code{log_tau2_2} (log-transformed nugget parameter for the Building Category 2 spatial field), \code{log_sigma2_2} (log-transformed partial sill parameter for the Building Category 2 spatial field), \code{log_slope1} (log-transformed slope parameter for Building Category 1) and \code{log_slope2} (log-transformed slope parameter for Building Category 2).}
\item{objective}{Final value of the objective function e.g. negative log-likelihood.}
\item{convergence}{Convergence code of the nlminb algorithm. 0 represents convergence.}
\item{iterations}{Number of iterations taken in optimisation procedure before stopping. Note that a maximum of 150 iterations have been set in the \code{spatial_ordinal} function.}
\item{evaluations}{Number of function and gradient evaluations taken in optimisation procedure before stopping.}
\item{message}{Message obtained from the nlminb algorithm.}
\item{field}{Estimated values of the field shared between different building categories with long-range spatial correlation.}
\item{field1}{Estimated values of the spatial field associated with building category 1 with shorter-range spatial correlation.}
\item{field2}{Estimated values of the spatial field associated with building category 2 with shorter-range spatial correlation.}
\item{par.sd}{Estimated standard deviation obtained from the \code{TMB::sdreport} function for the parameters in \code{par}.}
}
}
\source{
Obtained from a spatial ordinal model fit to the simulated damage for an earthquake event in Haiti.
}
\usage{
demo_spatial_fit
}
\description{
A list containing the output of \code{spatial_ordinal} when the function is applied to the damage_simulation data.
}
\keyword{datasets}
