% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_ses_list}
\alias{event_ses_list}
\title{Matching list for simulated OpenQuake events and their stochastic event sets of one year duration.}
\format{
A dataframe with 166882 rows and 2 variables:
\describe{
\item{event_id}{Event ID of a simulated earthquake event affecting the selected portfolio of buildings in \code{oq_sim} where the corresponding logPGA values are given in the column named after the event ID.}
\item{rlz_id}{Realisation ID indicating the year of occurrence of stochastic event set (SES) of one year duration. Note that we've considered a total of 1 million SES in our simulations.}
}
}
\source{
Simulated using the OpenQuake Engine.
}
\usage{
event_ses_list
}
\description{
A dataframe of OpenQuake event IDs and their corresponding realisation ID denoting the year of occurrence or stochastic event set.
}
\keyword{datasets}
