% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frag_curve.R
\name{frag_curve}
\alias{frag_curve}
\title{Compute fitted exceedance probabilities and plot fragility curves.}
\usage{
frag_curve(model, type = "ordinal", data, plot = TRUE)
}
\arguments{
\item{model}{An ordinal model fitted for \code{data} using \code{polr} from the R package \code{MASS} or a list of probit regression models fitted for the different damage states in \code{data} using \code{glm}. E.g. Grade 0 against Grades 1-5, Grades 0-1 against Grades 2-5 etc.}

\item{type}{"ordinal" (default) or "probit" if supplying a list of probit regressions for \code{model}.}

\item{data}{A dataframe with the columns \code{CDF} (ordered factor), \code{logPGA} (numeric), \code{PGA} (numeric), \code{Easting} (numeric) and \code{Northing} (numeric).}

\item{plot}{A logical value indicating if the fragility curves should be plotted.}
}
\value{
A dataframe with the columns \code{PGA}, \code{CDF} and \code{ex.prob} (the exceedance probability of the \code{CDF} level at the \code{PGA} value).
}
\description{
Compute fitted exceedance probabilities and plot fragility curves.
}
\examples{
library(MASS)

data(damage_simulation)

data.subset.1 <- damage_simulation[damage_simulation$building_cat == 1, ]

frag.model.1 <- polr(CDF ~ logPGA, data = data.subset.1,
                     method = "probit", Hess = TRUE)
ex.prob.1 <- frag_curve(frag.model.1, data = data.subset.1)
}
