% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragilitycurves-package.R
\docType{package}
\name{fragilitycurves-package}
\alias{fragilitycurves}
\alias{fragilitycurves-package}
\title{fragilitycurves: R package for developing fragility curves, particularly using ordinal models}
\description{
Functions and sample datasets for developing fragility curves. Particular focus is given to that built using ordinal regression and its extension to account for spatial correlation in damage.
}
\author{
\strong{Maintainer}: Michele Nguyen \email{michele.nguyen@ntu.edu.sg} (\href{https://orcid.org/0000-0001-9017-1978}{ORCID})

}
\keyword{internal}
