% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haiti_admin0}
\alias{haiti_admin0}
\title{Shapefile for the Haiti level 0, i.e. national administrative unit.}
\format{
SpatialPolygonsDataFrame containing 1 shapefile of the level 0, i.e. national administrative unit in Haiti.
}
\source{
Haiti Subnational Administrative Boundaries contributed by the UN office for the Coordination of Humanitarian Affairs (OCHA) country office in Haiti, available on the Humanitarian Data Exchange (HDX).
}
\usage{
haiti_admin0
}
\description{
A SpatialPolygonsDataFrame containing shapefiles of the level 0, i.e. national administrative unit in Haiti.
}
\keyword{datasets}
