% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haiti_admin2}
\alias{haiti_admin2}
\title{Shapefiles of Haiti level 2 administrative units.}
\format{
SpatialPolygonsDataFrame containing 140 shapefiles of level 2 administrative units in Haiti.
}
\source{
Haiti Subnational Administrative Boundaries contributed by the UN office for the Coordination of Humanitarian Affairs (OCHA) country office in Haiti, available on the Humanitarian Data Exchange (HDX).
}
\usage{
haiti_admin2
}
\description{
A SpatialPolygonsDataFrame containing shapefiles of level 2 administrative units in Haiti.
}
\keyword{datasets}
