% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriged_fields.R
\name{kriged_fields}
\alias{kriged_fields}
\title{Krige field estimates obtained via a spatial ordinal model fit.}
\usage{
kriged_fields(model.fit, data.1, data.2, zoom.shp = NULL, PGA.raster)
}
\arguments{
\item{model.fit}{The output of a \code{spatial_ordinal} model fit.}

\item{data.1}{The damage dataset of Building Category 1 used in the fit.}

\item{data.2}{The damage dataset of Building Category 2 used in the fit.}

\item{zoom.shp}{A shapefile of the area to zoom into and bound the kriging by (optional).}

\item{PGA.raster}{A raster of the peak ground acceleration (PGA) which covers \code{zoom.shp} if applicable and whose resolution determines that of the outputs'.}
}
\value{
A list of six rasters corresponding to the kriged shared field and building category specific fields together with their kriging variances.
}
\description{
Krige field estimates obtained via a spatial ordinal model fit.
}
\examples{
library(raster)
library(rgdal)
library(sp)

data(damage_simulation)
data(mean_PGA)
data(demo_spatial_fit)
data(haiti_admin2)

data.subset.1 <- damage_simulation[damage_simulation$building_cat == 1, ]
data.subset.2 <- damage_simulation[damage_simulation$building_cat == 2, ]

# Obtain the shapefile for Port-au-Prince:

pap_shp <- haiti_admin2[haiti_admin2$ADM2_EN == "Port-au-Prince", ]
pap_shp <- sp::spTransform(pap_shp, CRS("+proj=utm +zone=18 ellps=WGS84"))

kriged_rasters <- kriged_fields(demo_spatial_fit, data.subset.1,
                                data.subset.2, pap_shp, mean_PGA)
}
