% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_var.R
\name{latent_var}
\alias{latent_var}
\title{Compute the latent variable mean raster for a building category and plot its contributing terms.}
\usage{
latent_var(category, new.par, field.rasters, PGA.raster, study.shp = NULL)
}
\arguments{
\item{category}{Set this to 1 (Building category 1) or 2 (Building category 2).}

\item{new.par}{The vector with parameter estimates corresponding to a \code{spatial_ordinal} model fit or its equivalent with cut-off factors converted to the latent variable scale via \code{convert_cutoffs}.}

\item{field.rasters}{A list of rasters where the first three correspond to the kriged shared field and two building category specific fields respectively. This could be the output of \code{kriged_fields}.}

\item{PGA.raster}{A raster of the peak ground acceleration (PGA) which covers the extent of the rasters in \code{field.rasters}.}

\item{study.shp}{A shapefile of the study area which covers the area spanned by \code{PGA.raster} (optional). If not provided, the plot of the contributing terms will not be computed.}
}
\value{
A raster of the latent variable mean surface computed according to the fitted spatial ordinal model for the building category.
}
\description{
Compute the latent variable mean raster for a building category and plot its contributing terms.
}
\examples{
library(raster)
library(rgdal)
library(sp)
library(rgeos)

data(damage_simulation)
data(mean_PGA)
data(demo_spatial_fit)
data(haiti_admin2)
data(haiti_admin0)

# Convert the cut-off factors to cut-off values on the latent variable scale:
new_par <- convert_cutoffs(demo_spatial_fit$par)

study_shp <- sp::spTransform(haiti_admin0, sp::CRS("+proj=utm +zone=18 ellps=WGS84"))

# Obtain the shapefile for Port-au-Prince:
pap_shp <- haiti_admin2[haiti_admin2$ADM2_EN == "Port-au-Prince", ]
pap_shp <- sp::spTransform(pap_shp, sp::CRS("+proj=utm +zone=18 ellps=WGS84"))

data.subset.1 <- damage_simulation[damage_simulation$building_cat == 1, ]
data.subset.2 <- damage_simulation[damage_simulation$building_cat == 2, ]

kriged_rasters <- kriged_fields(demo_spatial_fit, data.subset.1,
                                data.subset.2, pap_shp, mean_PGA)
latent_var_1 <- latent_var(category = 1, new_par, kriged_rasters,
                           mean_PGA, study_shp)
}
