% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_exceed.R
\name{prob_exceed}
\alias{prob_exceed}
\title{Compute and plot the exceedance probability rasters for the different damage states.}
\usage{
prob_exceed(category, new.par, damage.states, latent.raster, study.shp)
}
\arguments{
\item{category}{Set this to 1 (Building category 1) or 2 (Building category 2).}

\item{new.par}{The vector with parameter estimates corresponding to a \code{spatial_ordinal} model fit or its equivalent with cut-off factors converted to the latent variable scale via \code{convert_cutoffs}.}

\item{damage.states}{A vector of the damage state names e.g. their central damage factor values (character or factor).}

\item{latent.raster}{A raster of the latent variable mean surface. This could be the output of the \code{latent_var} function.}

\item{study.shp}{A shapefile of the study area which covers the area spanned by \code{PGA.raster} (optional). If not provided, the plots of the exceedance probabilities will not be computed.}
}
\value{
A dataframe with columns corresponding to "Easting", "Northing" and the exceedance probabilities of the damage states.
}
\description{
Compute and plot the exceedance probability rasters for the different damage states.
}
\examples{
library(raster)
library(rgdal)
library(sp)
library(rgeos)

data(damage_simulation)
data(mean_PGA)
data(demo_spatial_fit)
data(haiti_admin2)
data(haiti_admin0)

# Convert the cut-off factors to cut-off values on the latent variable scale:
new_par <- convert_cutoffs(demo_spatial_fit$par)

study_shp <- sp::spTransform(haiti_admin0, CRS("+proj=utm +zone=18 ellps=WGS84"))

# Obtain the shapefile for Port-au-Prince:
pap_shp <- haiti_admin2[haiti_admin2 == "Port-au-Prince", ]
pap_shp <- sp::spTransform(pap_shp, CRS("+proj=utm +zone=18 ellps=WGS84"))

data.subset.1 <- damage_simulation[damage_simulation$building_cat == 1, ]

kriged_rasters <- kriged_fields(demo_spatial_fit, data.subset.1,
                                data.subset.2, pap_shp, mean_PGA)
latent_var_1 <- latent_var(category = 1, new_par, kriged_rasters,
                           mean_PGA, study_shp)

CDF_breaks <- sort(unique(data.subset.1$CDF), decreasing = FALSE)

exceed_prob_1 <- prob_exceed(1, new_par, CDF_breaks, latent_var_1, study_shp)
}
