% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel_aic.R
\name{submodel_aic}
\alias{submodel_aic}
\title{Compute Akaike Information Criterion (AIC) values for the Damage-spatial model and its Non-spatial and IM-spatial submodels.}
\usage{
submodel_aic(data.1, data.2, model.fit)
}
\arguments{
\item{data.1}{The damage dataset of Building Category 1 used in the fit.}

\item{data.2}{The damage dataset of Building Category 2 used in the fit.}

\item{model.fit}{The output of a \code{spatial_ordinal} model fit.}
}
\value{
A vector containing the AIC values for the Non-spatial and IM-spatial submodels as well as the Damage-spatial model.
}
\description{
Compute Akaike Information Criterion (AIC) values for the Damage-spatial model and its Non-spatial and IM-spatial submodels.
}
\examples{
data(demo_spatial_fit)
data(damage_simulation)

data.subset.1 <- damage_simulation[damage_simulation$building_cat == 1, ]
data.subset.2 <- damage_simulation[damage_simulation$building_cat == 2, ]

# The following takes 10 seconds on a PC with characteristics:
# Intel(R) Xeon (R) W-2112 CPU Processor @ 3.60GHz; 32GB of RAM;
aic_val <- submodel_aic(data.1 = data.subset.1,
                         data.2 = data.subset.2, model.fit = demo_spatial_fit)
}
